Add header component with links and routes to app component
