+++ src/views/blog.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Blog() {
+  return (
+    <div><h1>blog</h1></div>
+  )
+}
+
+export default Blog
